requirements.txt
# Proratio - AI-Driven Crypto Trading System
# Requirements file for all modules

# ============================================================================
# CORE FRAMEWORK
# ============================================================================
freqtrade==2025.9.1
ccxt==4.5.8

# ============================================================================
# DATABASE
# ============================================================================
psycopg2-binary==2.9.10
sqlalchemy==2.0.43

# ============================================================================
# AI / LLM PROVIDERS
# ============================================================================
openai==2.2.0
anthropic==0.69.0
google-generativeai==0.8.5

# ============================================================================
# DATA & ANALYSIS
# ============================================================================
pandas==2.3.3
numpy==2.3.3
ft-pandas-ta==0.3.16
TA-Lib==0.6.7
# TA-Lib requires separate binary installation:
# Mac: brew install ta-lib
# Linux: See https://github.com/mrjbq7/ta-lib

# ============================================================================
# MACHINE LEARNING (FreqAI & Models)
# ============================================================================
# Note: PyTorch is large (~2GB). Installed for LSTM models.
torch==2.8.0                   # Deep learning (LSTM/GRU models)
torchvision                    # Computer vision utilities
torchaudio                     # Audio processing utilities
scikit-learn==1.7.2
scipy==1.16.2
lightgbm==4.5.0                # FreqAI gradient boosting
xgboost==2.1.3                 # FreqAI alternative boosting model
catboost==1.2.7                # FreqAI alternative boosting model
joblib==1.4.2                  # Model serialization
optuna==4.1.0                  # Hyperparameter optimization
shap==0.47.0                   # Model interpretability

# ============================================================================
# VISUALIZATION & DASHBOARDS
# ============================================================================
plotly==6.3.1
matplotlib==3.10.6
streamlit==1.50.0
seaborn==0.13.2

# ============================================================================
# CLI & INTERFACE
# ============================================================================
typer==0.19.2                  # CLI framework
rich==13.9.4                   # Rich terminal output
shellingham==1.5.4             # Shell detection for completion

# ============================================================================
# UTILITIES
# ============================================================================
python-dotenv==1.1.1
requests==2.32.5
pydantic==2.12.0
pydantic-settings==2.11.0
python-telegram-bot==22.5
fastapi==0.118.0
uvicorn==0.37.0

# ============================================================================
# DEVELOPMENT & TESTING
# ============================================================================
pytest==8.4.2
pytest-asyncio==1.2.0
pytest-cov==7.0.0
black==25.9.0
mypy==1.18.2

# ============================================================================
# JUPYTER & RESEARCH
# ============================================================================
jupyterlab==4.4.9
